feat: add database connection and shutdown functionality
