+++ api/database.py
@@ -3,20 +3,27 @@ from typing import Annotated
 
 from dotenv import load_dotenv
 from fastapi import Depends
+from databases import Database
 from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker, Session
+from utils.logger import logger
 
 load_dotenv('.env')
 
 # Define the URI for the postgres database
 DATABASE_URI = os.getenv('DATABASE_URI')
 
+if not DATABASE_URI:
+    logger.error("DATABASE_URI is not set in the environment variables.")
+    raise ValueError("DATABASE_URI is not set in the environment variables.")
+
 # Create the database engine and session
 try:
     engine = create_engine(DATABASE_URI)
+    database = Database(DATABASE_URI)
 except Exception as e:
-    print(f"Failed to create engine: {e}")
+    logger.error(f"Failed to create engine: {e}")
 
 # Create a session local
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
